+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Date format: DD/MM/YYYY
 
 ## [3.5.1] - [DD/MM/YYYY]
 
+- Add the `borderRadius` and `shape` attributes to the `Mica` widget
 - Implement `DropDownButton` ([#85](https://github.com/bdlukaa/fluent_ui/issues/85))
 
 ## [3.5.0] - Flutter 2.8 - [09/12/2021]